XCOPY
I created this program to save myself a lot of typing on the command
line when I was asked to backup a bunch of the archive data stored on
BD-R disk.
The user simply inputs the destination pathname on the configuration
page first, then the program refreshes itself with saved user settings.
The user can then select the destination name and press start to
activate xcopy.